Use the Release API instead of checkpoint.
