+++ example/terraform.pkr.hcl
@@ -1,7 +1,7 @@
 packer {
   required_plugins {
     terraform = {
-      version = ">= 0.0.7"
+      version = ">= 0.0.8"
       source  = "github.com/servian/terraform"
     }
   }
